+++ CHANGELOG.md
@@ -15,6 +15,8 @@
   * presentRating(Context, String)
   * presentRating(Context, String, FeedbackCallback)
 
+* Mitigated an issue where content communication was done twice.
+
 ## 24.7.4
 * Disabled caching for webviews.
 * Expanded the flag (enablePreviousNameRecording) to add current view name as segmentation to custom events. (Experimental!)